requirements_vercel.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
yfinance==0.2.28
plotly==5.18.0
fredapi==0.5.1
openai==1.12.0
newsapi-python==0.2.6
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0 